# Allow concatenating small InlineStrings to return an InlineString