--- a/src/InlineStrings.jl
+++ b/src/InlineStrings.jl
@@ -620,8 +620,14 @@ Base.string(a::BaseStrs, b::BaseStrs, c::InlineString) = _string(a, b, c)
 end
 
 # For more and/or bigger InlineStrings creating a `Base.String` is faster
-const TinyInlineStrings = Union{InlineString1, InlineString3, InlineString7}
-Base.string(a::TinyInlineStrings, b::TinyInlineStrings, c::TinyInlineStrings) = _string(_string(a, b), c)
+const _SmallerInlineStrings = Union{InlineString1, InlineString3, InlineString7}
+Base.string(a::_SmallerInlineStrings, b::_SmallerInlineStrings, c::_SmallerInlineStrings) =
+    _string(_string(a, b), c)
+const _SmallestInlineStrings = Union{InlineString1, InlineString3}
+Base.string(a::_SmallestInlineStrings, b::_SmallestInlineStrings, c::_SmallestInlineStrings) =
+    _string(_string(a, b), c)
+Base.string(a::_SmallestInlineStrings, b::_SmallestInlineStrings, c::_SmallestInlineStrings, d::_SmallestInlineStrings) =
+    _string(_string(_string(a, b), c), d)
 
 # Only benefit from keeping the small-ish strings as InlineStrings
 function _string(a::Ta, b::Tb) where {Ta <: SmallInlineStrings, Tb <: SmallInlineStrings}