- closes #58 
- now string concatenation `*` / `string` return an `InlineString` in some cases which give a performance improvment
    - the `Base.*` / `Base.string` API does not require us to return a `Base.String` (just an `AbstractString`)
        ```julia
        help?> *
        search: *
        
          *(s::Union{AbstractString, AbstractChar}, t::Union{AbstractString, AbstractChar}...) -> AbstractString
        
          Concatenate strings and/or characters, producing a String. This is equivalent to calling the string function on the
          arguments. Concatenation of built-in string types always produces a value of type String but other string types may choose
          to return a string of a different type as appropriate.
        ```

### Micro-benchmarks

```julia
using InlineStrings
using BenchmarkTools

s1 = InlineString1("a")
s3 = InlineString3("bb")
s7 = InlineString7("ccc")
s15 = InlineString15("dddd")

for s in (s1, s3, s7, s15)
    println(typeof(s))
    @btime $s * $s
    @btime $s * $s * $s
    @btime $s * $s * $s * $s
    @btime $s * $s * $s * $s * $s
end
```

#### Now (this PR)
```julia
String1
  2.125 ns (0 allocations: 0 bytes)
  2.125 ns (0 allocations: 0 bytes)
  2.084 ns (0 allocations: 0 bytes)
  7.125 ns (1 allocation: 24 bytes)
String3
  2.083 ns (0 allocations: 0 bytes)
  2.416 ns (0 allocations: 0 bytes)
  5.500 ns (0 allocations: 0 bytes)
  14.822 ns (1 allocation: 32 bytes)
String7
  2.416 ns (0 allocations: 0 bytes)
  4.500 ns (0 allocations: 0 bytes)
  19.163 ns (1 allocation: 32 bytes)
  22.234 ns (1 allocation: 32 bytes)
String15
  3.958 ns (0 allocations: 0 bytes)
  16.992 ns (1 allocation: 32 bytes)
  20.436 ns (1 allocation: 40 bytes)
  23.803 ns (1 allocation: 40 bytes)
```
#### Before (v1.3.2)
```julia
String1
  6.791 ns (1 allocation: 24 bytes)
  6.958 ns (1 allocation: 24 bytes)
  7.083 ns (1 allocation: 24 bytes)
  40.825 ns (6 allocations: 104 bytes)
String3
  9.844 ns (1 allocation: 24 bytes)
  11.678 ns (1 allocation: 24 bytes)
  13.263 ns (1 allocation: 32 bytes)
  48.963 ns (6 allocations: 112 bytes)
String7
  12.638 ns (1 allocation: 24 bytes)
  16.073 ns (1 allocation: 32 bytes)
  19.163 ns (1 allocation: 32 bytes)
  56.572 ns (6 allocations: 112 bytes)
String15
  13.597 ns (1 allocation: 32 bytes)
  16.951 ns (1 allocation: 32 bytes)
  20.416 ns (1 allocation: 40 bytes)
  59.070 ns (6 allocations: 200 bytes)
```

Run on v1.8.4:
```julia
julia> versioninfo()
Julia Version 1.8.4
Commit 00177ebc4fc (2022-12-23 21:32 UTC)
Platform Info:
  OS: macOS (arm64-apple-darwin21.5.0)
  CPU: 10 × Apple M1 Max
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-13.0.1 (ORCJIT, apple-m1)
  Threads: 1 on 8 virtual cores
```